backend/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# SSH Automation
paramiko>=3.1.0
fabric>=3.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
# psycopg2-binary>=2.9.5  # For PostgreSQL - Uncomment if using PostgreSQL
aiosqlite>=0.19.0       # For SQLite

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Scheduling
python-crontab>=2.7.1
croniter>=1.3.8

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
coverage>=7.2.0

# Utilities
pyyaml>=6.0.0
loguru>=0.7.0 